refactor: Extract CLI test helper functions into their own module
